Add files via upload
For this project, I wanted to create something to memorialize one of my favorite albums. Knowing the original object, the 2006 album by The Killers, I was able to begin from that basic visual vocabulary. My original image chosen is from the actual casino in Las Vegas that the album is named after. By desaturating this image, I was given the aesthetic of old Las Vegas I was trying to achieve. From there, I knew I wanted to play with the images on the projection screen on the casino sign. I had prior experience with GIMP, so I knew that I could create a different image to superimpose on to the part of the sign I desired. It was the most difficult to get rid of the original image projected on the sign, and see how I could convincingly put a different image in its place. It was also difficult to make sure that the text on the sign was readable and relevant. In my original draft, the text I used on the sign was kind of jarring and did not fit with the branding I was trying to achieve. I am satisfied with my final draft and I think it fits the desired effect I was trying to achieve. My added text is readable and concise and I think that the screen image looks authentic. I think I ended up with a a strong and simple image that conveys the message I wanted.